Extract useIsTruncated to a separate file
