+++ frontend/src/metabase/hooks/use-is-truncated.ts
@@ -1,0 +1,36 @@
+import { useLayoutEffect, useRef, useState } from "react";
+
+import resizeObserver from "metabase/lib/resize-observer";
+
+export const useIsTruncated = <E extends HTMLElement>() => {
+  const ref = useRef<E | null>(null);
+  const [isTruncated, setIsTruncated] = useState(false);
+
+  useLayoutEffect(() => {
+    const element = ref.current;
+
+    if (!element) {
+      return;
+    }
+
+    const handleResize = () => {
+      setIsTruncated(getIsTruncated(element));
+    };
+
+    handleResize();
+    resizeObserver.subscribe(element, handleResize);
+
+    return () => {
+      resizeObserver.unsubscribe(element, handleResize);
+    };
+  }, []);
+
+  return { isTruncated, ref };
+};
+
+const getIsTruncated = (element: HTMLElement): boolean => {
+  return (
+    element.scrollHeight > element.clientHeight ||
+    element.scrollWidth > element.offsetWidth
+  );
+};
